refactor(client): move the de facto shared chat components into chat-primitives
Sixteen components the DM screen already imported straight out of
channel-view/text (reactions, file card, media players, image/link/
code/social overrides, composer expand, formatting hints, reply
content preview, error boundary) move to components/chat-primitives —
pure renames plus import rewrites across nine consumers, no logic
changes. First tranche of the chat-stack deduplication.

Also fixes two federated id-space bugs in surfaces touched by the
move: the DM pin-banner author and DM search results resolved
HOME-space user ids against the ambient roster, showing the wrong
name/avatar whenever a federated server was active.
