+++ apps/client/src/components/channel-view/text/renderer/index.tsx
@@ -25,13 +25,13 @@ import parse from 'html-react-parser';
 import { memo, useCallback, useMemo, useState } from 'react';
 import { toast } from 'sonner';
 import { Tooltip } from '../../../ui/tooltip';
-import { FileCard } from '../file-card';
-import { ImageContextMenu } from '../image-context-menu';
-import { MessageReactions } from '../message-reactions';
-import { AudioPlayer } from '../overrides/audio-player';
-import { ImageOverride } from '../overrides/image';
-import { LinkPreview } from '../overrides/link-preview';
-import { VideoPlayer } from '../overrides/video-player';
+import { FileCard } from '@/components/chat-primitives/file-card';
+import { ImageContextMenu } from '@/components/chat-primitives/image-context-menu';
+import { MessageReactions } from '@/components/chat-primitives/message-reactions';
+import { AudioPlayer } from '@/components/chat-primitives/overrides/audio-player';
+import { ImageOverride } from '@/components/chat-primitives/overrides/image';
+import { LinkPreview } from '@/components/chat-primitives/overrides/link-preview';
+import { VideoPlayer } from '@/components/chat-primitives/overrides/video-player';
 import { serializer } from './serializer';
 import type { TFoundMedia } from './types';
