+++ apps/client/src/components/chat-primitives/composer-expand.tsx
