+++ apps/client/src/components/chat-primitives/file-card.tsx
