+++ apps/client/src/components/chat-primitives/formatting-hints.tsx
