+++ apps/client/src/components/chat-primitives/image-context-menu.tsx
