+++ apps/client/src/components/chat-primitives/message-error-boundary.tsx
