+++ apps/client/src/components/chat-primitives/overrides/image.tsx
