+++ apps/client/src/components/chat-primitives/overrides/layout.tsx
